Enable+apply more lints

+++ src/OpenSimCreator/Widgets/ObjectPropertiesEditor.cpp
@@ -34,6 +34,7 @@
 
 #include <algorithm>
 #include <array>
+#include <cstddef>
 #include <memory>
 #include <string>
 #include <type_traits>
@@ -864,7 +865,7 @@ namespace
                 ImGui::PushID(i);
 
                 ImGui::SetNextItemWidth(ImGui::GetContentRegionAvail().x);
-                if (ImGui::InputFloat3("##vec6editor", rawValue.data() + 3*i, "%.6f"))
+                if (ImGui::InputFloat3("##vec6editor", rawValue.data() + static_cast<ptrdiff_t>(3*i), "%.6f"))
                 {
                     m_EditedProperty.updValue(idx)[3*i + 0] = static_cast<double>(rawValue[3*i + 0]);
                     m_EditedProperty.updValue(idx)[3*i + 1] = static_cast<double>(rawValue[3*i + 1]);
